Adds protocol tests and updates dependencies
Adds a new test project for the AvantiPoint.Packages.Protocol library.
This project includes tests for download, metadata, and search functionalities.
These tests ensure the correct behavior of the NuGet client when interacting with a NuGet feed.

Additionally, this commit updates the Visual Studio version and several package dependencies to their latest versions.
This ensures compatibility and access to the latest features and bug fixes.